py-backend-api/requirements.txt
google-cloud-aiplatform==1.43.0
Flask==2.3.2
Flask-Cors==3.0.10
urllib3==1.26.15
sqlglot==18.17.0
langchain==0.0.323
itables==1.6.2
langchain-experimental==0.0.34
pydantic==2.4.2
sqlalchemy-bigquery==1.8.0
SQLAlchemy==1.4.49
# google-cloud-aiplatform==1.35.0
gradio==3.50.2
pygwalker==0.3.9
google-cloud-bigquery
cmake
pandas-gbq==0.19.2
python-dotenv==1.0.1
